moved applyPrecession and applyProperMotion to the new AstrometryUtils.py which defines functions that are not associated with a class
