+++ python/lsst/sims/coordUtils/__init__.py
@@ -1,1 +1,2 @@
 from .Astrometry import *
+from .AstrometryUtils import *
